requirements.txt
django_tinymce4_lite==1.8.0
Django==3.0.5
